+++ apps/dashboard/app/(app)/apis/[apiId]/settings/actions.ts
@@ -1,13 +1,7 @@
 "use server";
 
-import { revalidatePath, revalidateTag } from "next/cache";
+import { revalidatePath } from "next/cache";
 
 async function revalidate() {
   await revalidatePath("/", "layout");
 }
-
-async function revalidateMyTag(tag: string) {
-  revalidateTag(tag);
-}
-
-export { revalidate, revalidateMyTag as revalidateTag };
