+++ apps/dashboard/app/(app)/apis/[apiId]/settings/update-api-name.tsx
@@ -17,7 +17,7 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useRouter } from "next/navigation";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
-import { revalidateTag } from "./actions";
+import { revalidateTag } from "../../../../actions";
 const formSchema = z.object({
   name: z.string(),
   apiId: z.string(),
