+++ apps/dashboard/app/(app)/authorization/permissions/[permissionId]/client.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { revalidateTag } from "@/app/(app)/apis/[apiId]/settings/actions";
+import { revalidateTag } from "@/app/actions";
 import { Loading } from "@/components/dashboard/loading";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardFooter } from "@/components/ui/card";
