+++ apps/dashboard/app/(app)/authorization/permissions/page.tsx
@@ -43,6 +43,7 @@ export default async function RolesPage() {
     return redirect("/new");
   }
 
+  console.log(workspace.permissions, `workspace.permissions`)
   /**
    * Filter out all the soft deleted keys cause I'm not smart enough to do it with drizzle
    */
