+++ apps/dashboard/app/(app)/ratelimits/[namespaceId]/page.tsx
@@ -19,7 +19,7 @@ import {
 } from "@/lib/tinybird";
 import { BarChart } from "lucide-react";
 import ms from "ms";
-import { notFound } from "next/navigation";
+import { notFound, redirect } from "next/navigation";
 import { parseAsArrayOf, parseAsString, parseAsStringEnum } from "nuqs/server";
 import { Filters, type Interval } from "./filters";
 
@@ -49,7 +49,8 @@ export default async function RatelimitNamespacePage(props: {
     },
   });
   if (!namespace || namespace.workspace.tenantId !== tenantId) {
-    return notFound();
+    redirect("/ratelimits");
+    return;
   }
 
   const interval = intervalParser.withDefault("7d").parseServerSide(props.searchParams.interval);
