+++ apps/dashboard/app/(app)/ratelimits/page.tsx
@@ -37,6 +37,8 @@ export default async function RatelimitOverviewPage() {
     return redirect("/new");
   }
 
+  console.log(workspace.ratelimitNamespaces,`workspace.ratelimitNamespaces`)
+
   const snippet = `curl -XPOST 'https://api.unkey.dev/v1/ratelimits.limit' \\
   -H 'Content-Type: application/json' \\
   -H 'Authorization: Bearer <UNKEY_ROOT_KEY>' \\
