+++ apps/dashboard/app/actions.ts
@@ -1,6 +1,12 @@
 "use server";
-import { revalidatePath } from "next/cache";
+import { revalidatePath, revalidateTag } from "next/cache";
 
 export async function revalidate(path: string) {
   revalidatePath(path, "page");
 }
+
+export async function revalidateMyTag(slug: string) {
+  revalidateTag(slug)
+}
+
+export { revalidateMyTag as revalidateTag };
